rename: 위젯 디렉토리 이름 변경 (사이드 바 -> 위젯 바)
