+++ src/components/common/WidgetsSection.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import SearchBar from 'components/ui/SearchBar'
-import AnnouncementWidget from 'components/community/side-bar/AnnouncementWidget'
-import BestPostsWidget from 'components/community/side-bar/BestPostsWidget'
-import AdvertisementWidget from 'components/community/side-bar/AdvertisementWidget'
-import CommentRankWidget from 'components/community/side-bar/CommentRankWidget'
-import GatherPeopleWidget from 'components/community/side-bar/GatherPeopleWidget'
+import AnnouncementWidget from 'components/community/widget-bar/AnnouncementWidget'
+import BestPostsWidget from 'components/community/widget-bar/BestPostsWidget'
+import AdvertisementWidget from 'components/community/widget-bar/AdvertisementWidget'
+import CommentRankWidget from 'components/community/widget-bar/CommentRankWidget'
+import GatherPeopleWidget from 'components/community/widget-bar/GatherPeopleWidget'
 
 const WidgetsSection = () => {
   return (
