--- src/components/community/side-bar/AdvertisementWidget.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import { ShadowBox } from 'styles/reusable-style/elementStyle'
-
-const AdvertisementWidget = () => {
-  return <ShadowBox>AdvertisementWidget</ShadowBox>
-}
-
-export default AdvertisementWidget
