+++ src/components/community/widget-bar/AnnouncementWidget.tsx
