+++ src/components/community/widget-bar/GatherPeopleWidget.tsx
